langchanger/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
decorator==4.4.2
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.2.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-jsonapi==3.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
future==0.18.2
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
